+++ src/main/java/com/crypto/portfolio/web/UserController.java
@@ -5,10 +5,13 @@ import com.crypto.portfolio.services.UserService;
 
 import org.springframework.http.HttpStatus;
 
+import java.util.List;
 import java.util.Optional;
-import java.lang.Iterable;
+import java.util.stream.Collectors;
+import java.util.stream.StreamSupport;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
@@ -23,52 +26,71 @@ public class UserController {
     }
 
     @GetMapping("/{id}")
-    public Optional<User> getHolder(@PathVariable Integer id) {
-        return service.getUserById(id);
+    public ResponseEntity<?> getUserById(@PathVariable Integer id) {
+        Optional<User> user = service.getUserById(id);
+        if (user.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("UserNotFound");
+        }
+        return ResponseEntity.ok(user);
     }
 
     @GetMapping("/telegramId/{telegramId}")
-    public Optional<User> getHolderByTelegramId(@PathVariable Long telegramId) {
-        return service.getUserByTelegramId(telegramId);
+    public ResponseEntity<?> getUserByTelegramId(@PathVariable Long telegramId) {
+        Optional<User> user = service.getUserByTelegramId(telegramId);
+        if (user.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("UserNotFound");
+        }
+        return ResponseEntity.ok(user);
     }
 
     @GetMapping("/userName/{userName}")
-    public Optional<User> getHolderByUserName(@PathVariable String userName) {
-        return service.getUserByUserName(userName);
+    public ResponseEntity<?> getUserByUserName(@PathVariable String userName) {
+        Optional<User> user = service.getUserByUserName(userName);
+        if (user.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("UserNotFound");
+        }
+        return ResponseEntity.ok(user);
     }
 
     @PostMapping("/add")
-    @ResponseStatus(HttpStatus.CREATED)
-    public void createHolder(@RequestBody User user) {
-        service.saveUser(user);
+    public ResponseEntity<?> addUser(@RequestBody User user) {
+        if (!service.getUserByTelegramId(user.getTelegramId()).isEmpty()) {
+            return ResponseEntity.status(HttpStatus.CONFLICT).body("The user already exists.");
+        }
+        User savedUser = service.saveUser(user);
+        return ResponseEntity.status(HttpStatus.CREATED).body(user);
     }
 
     @PutMapping("/update/{id}")
-    @ResponseStatus(HttpStatus.ACCEPTED)
-    public User updateHolder(@PathVariable Integer id, @RequestBody User user) {
-        User userObject = null;
-        try {
-            userObject = (service.getUserById(id)).get();
-            userObject.setName(user.getName());
-            userObject.setTelegramId(user.getTelegramId());
-            userObject.setTelegramUserName(user.getTelegramUserName());
-            userObject.setComment(user.getComment());
-            userObject.setUserRole(user.getUserRole());
-            userObject.setTelegramGroupId(user.getTelegramGroupId());
-        } catch (Exception e) {
-            System.err.println(e.getStackTrace());
+    public ResponseEntity<?> updateHolder(@PathVariable Integer id, @RequestBody User user) {
+        Optional<User> userById = service.getUserById(id);
+        if (userById.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("UserNotFound");
         }
-        return service.saveUser(userObject);
+        User existingUser = userById.get();
+        existingUser.setName(user.getName());
+        existingUser.setTelegramId(user.getTelegramId());
+        existingUser.setTelegramUserName(user.getTelegramUserName());
+        existingUser.setComment(user.getComment());
+        existingUser.setUserRole(user.getUserRole());
+        existingUser.setTelegramGroupId(user.getTelegramGroupId());
+        service.saveUser(existingUser);
+        return ResponseEntity.ok(existingUser);
     }
 
     @GetMapping
-    public Iterable<User> getAllHolders(){
-        return service.getAllUsers();
+    public ResponseEntity<?> getAllHolders() {
+        List<User> result = StreamSupport.stream(service.getAllUsers().spliterator(), false)
+                        .collect(Collectors.toList());
+        return ResponseEntity.ok(result);
     }
 
     @DeleteMapping("/delete/{id}")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteHolder(@PathVariable Integer id) {
+    public ResponseEntity<?> deleteHolder(@PathVariable Integer id) {
+        if (service.getUserById(id).isEmpty()){
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
         service.deleteUserById(id);
+        return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
     }
 }
